+++ src/renderer/components/Settings/SettingsContext.ts
@@ -5,7 +5,6 @@ interface SettingsContextType {
 	// Global settings
 	title: string
 	selectConfigAtLaunch: boolean
-	setSelectConfigAtLaunch: (value: boolean) => void
 
 	// Auto-open at launch
 	autoOpenSelector: boolean